MDL-23271 - missing completion on submit for feedback

+++ mod/feedback/lib.php
@@ -28,6 +28,8 @@
 require_once($CFG->libdir.'/eventslib.php');
 /** Include calendar/lib.php */
 require_once($CFG->dirroot.'/calendar/lib.php');
+/** Include completionlib.php */
+require_once($CFG->libdir . '/completionlib.php');
 
 
 define('FEEDBACK_ANONYMOUS_YES', 1);
@@ -58,6 +60,7 @@ function feedback_supports($feature) {
         case FEATURE_GROUPMEMBERSONLY:        return true;
         case FEATURE_MOD_INTRO:               return true;
         case FEATURE_COMPLETION_TRACKS_VIEWS: return true;
+        case FEATURE_COMPLETION_HAS_RULES:    return true;
         case FEATURE_GRADE_HAS_GRADE:         return false;
         case FEATURE_GRADE_OUTCOMES:          return false;
         case FEATURE_BACKUP_MOODLE2:          return true;
@@ -417,6 +420,31 @@ function feedback_print_recent_mod_activity($activity, $courseid, $detail, $modn
     echo "</td></tr></table>";
 
     return;
+}
+
+/**
+ * Obtains the automatic completion state for this feedback based on the condition
+ * in feedback settings.
+ *
+ * @param object $course Course
+ * @param object $cm Course-module
+ * @param int $userid User ID
+ * @param bool $type Type of comparison (or/and; can be used as return value if no conditions)
+ * @return bool True if completed, false if not, $type if conditions not set.
+ */
+function feedback_get_completion_state($course, $cm, $userid, $type) {
+    global $CFG, $DB;
+
+    // Get feedback details
+    $feedback = $DB->get_record('feedback', array('id'=>$cm->instance), '*', MUST_EXIST);
+
+    // If completion option is enabled, evaluate it and return true/false 
+    if($feedback->completionsubmit) {
+        return $DB->record_exists('feedback_tracking', array('userid'=>$userid, 'feedback'=>$feedback->id));
+    } else {
+        // Completion option is not enabled so just return $type
+        return $type;
+    }
 }
 
 
@@ -1037,6 +1065,7 @@ function feedback_items_from_template($feedback, $templateid, $deleteold = false
     //files in the template_item are in the context of the current course
     //files in the feedback_item are in the feedback_context of the feedback
     $c_context = get_context_instance(CONTEXT_COURSE, $feedback->course);
+    $course = $DB->get_record('course', array('id'=>$feedback->course));
     $cm = get_coursemodule_from_instance('feedback', $feedback->id);
     $f_context = get_context_instance(CONTEXT_MODULE, $cm->id);
 
@@ -1050,7 +1079,17 @@ function feedback_items_from_template($feedback, $templateid, $deleteold = false
             }
             //delete tracking-data
             $DB->delete_records('feedback_tracking', array('feedback'=>$feedback->id));
-            $DB->delete_records('feedback_completed', array('feedback'=>$feedback->id));
+            
+            if($completeds = $DB->get_records('feedback_completed', array('feedback'=>$feedback->id))) {
+                $completion = new completion_info($course);
+                foreach($completeds as $completed) {
+                    // Update completion state
+                    if ($completion->is_enabled($cm) && $feedback->completionsubmit) {
+                        $completion->update_state($cm, COMPLETION_INCOMPLETE, $completed->userid);
+                    }
+                    $DB->delete_records('feedback_completed', array('id'=>$completed->id));
+                }
+            }
             $DB->delete_records('feedback_completedtmp', array('feedback'=>$feedback->id));
         }
         $positionoffset = 0;
@@ -1331,14 +1370,38 @@ function feedback_delete_item($itemid, $renumber = true){
 function feedback_delete_all_items($feedbackid){
     global $DB;
 
+    if(!$feedback = $DB->get_record('feedback', array('id'=>$feedbackid))) {
+        return false;
+    }
+    
+    if (!$cm = get_coursemodule_from_instance('feedback', $feedback->id)) {
+        return false;
+    }
+    
+    if(!$course = $DB->get_record('course', array('id'=>$feedback->course))) {
+        return false;
+    }
+    
     if(!$items = $DB->get_records('feedback_item', array('feedback'=>$feedbackid))) {
         return;
     }
     foreach($items as $item) {
         feedback_delete_item($item->id, false);
     }
+    // $DB->delete_records('feedback_completed', array('feedback'=>$feedbackid));
+    if($completeds = $DB->get_records('feedback_completed', array('feedback'=>$feedback->id))) {
+        $completion = new completion_info($course);
+        foreach($completeds as $completed) {
+            // Update completion state
+            if ($completion->is_enabled($cm) && $feedback->completionsubmit) {
+                $completion->update_state($cm, COMPLETION_INCOMPLETE, $completed->userid);
+            }
+            $DB->delete_records('feedback_completed', array('id'=>$completed->id));
+        }
+    }
+    
     $DB->delete_records('feedback_completedtmp', array('feedback'=>$feedbackid));
-    $DB->delete_records('feedback_completed', array('feedback'=>$feedbackid));
+    
 }
 
 /**
@@ -2275,6 +2338,19 @@ function feedback_delete_completed($completedid) {
     if (!$completed = $DB->get_record('feedback_completed', array('id'=>$completedid))) {
         return false;
     }
+    
+    if (!$feedback = $DB->get_record('feedback', array('id'=>$completed->feedback))) {
+        return false;
+    }
+    
+    if (!$course = $DB->get_record('course', array('id'=>$feedback->course))) {
+        return false;
+    }
+    
+    if (!$cm = get_coursemodule_from_instance('feedback', $feedback->id)) {
+        return false;
+    }
+
     //first we delete all related values
     $DB->delete_records('feedback_value', array('completed'=>$completed->id));
 
@@ -2283,6 +2359,11 @@ function feedback_delete_completed($completedid) {
         $DB->delete_records('feedback_tracking', array('completed'=>$completed->id));
     }
 
+    // Update completion state
+    $completion = new completion_info($course);
+    if ($completion->is_enabled($cm) && $feedback->completionsubmit) {
+        $completion->update_state($cm, COMPLETION_INCOMPLETE, $completed->userid);
+    }
     //last we delete the completed-record
     return $DB->delete_records('feedback_completed', array('id'=>$completed->id));
 }
